promote PR ID and line number to int64
Fixes #2443

PR ID and line number are were previously downcasted to int32 in a
number of places despite being mostly treated as an int64 type. While
the likelihood of either of these values exceeding a 32-bit MAX_INT are
minimal, this change makes them 64-bit to err on the side of caution.

+++ internal/engine/ingester/diff/diff.go
@@ -216,7 +216,7 @@ func ingestFileForFullDiff(filename, patch, patchUrl string) (*pb.PrContents_Fil
 		} else if strings.HasPrefix(line, "+") {
 			result = append(result, &pb.PrContents_File_Line{
 				Content:    line[1:],
-				LineNumber: int32(currentLineNumber),
+				LineNumber: currentLineNumber,
 			})
 
 			currentLineNumber++
